feat: add test_load_statistics_page_select_timespan with related methods, plus related refactoring
