+++ pages/manage_jenkins/status_information/system_information_page.py
@@ -4,7 +4,7 @@ from selenium.common import TimeoutException
 from selenium.webdriver.support.select import Select
 from selenium.webdriver.remote.webelement import WebElement
 from pages.manage_jenkins.manage_jenkins_page import ManageJenkinsPage
-from tests.manage_jenkins.status_information.data import SystemInformationData as SI
+from tests.manage_jenkins.data import SystemInformationData as SI
 
 
 class SystemInformationPage(ManageJenkinsPage):
@@ -13,7 +13,6 @@ class SystemInformationPage(ManageJenkinsPage):
         ACTIVE_TAB = (By.XPATH, "//div[@class='jenkins-tab-pane'][@style='display: block;']/h2")
         TABLE_ROWS = (By.TAG_NAME, "tr")
         TABLE_CELLS = (By.TAG_NAME, "td")
-        TIMESPAN_DROPDOWN = (By.ID, "timespan-select")
         THREAD_DUMP_PAGE_LINK = (By.XPATH, "//a[@href='threadDump']")
 
         @staticmethod
@@ -39,10 +38,6 @@ class SystemInformationPage(ManageJenkinsPage):
         @staticmethod
         def tab_by_name(tab_name: str) -> tuple[str, str]:
             return By.XPATH, f"//a[@href='#'][text()='{tab_name}']/.."
-
-        @staticmethod
-        def memory_usage_graph(flag_value: str) -> tuple[str, str]:
-            return By.XPATH, f"//img[@alt='Memory usage graph' and contains(@src, '{flag_value}')]"
 
     def __init__(self, driver, timeout=5):
         super().__init__(driver, timeout=timeout)
@@ -102,18 +97,6 @@ class SystemInformationPage(ManageJenkinsPage):
 
     def click_on_thread_dumps_tab(self) -> None:
         self.click_on_tab(SI.TABS_BAR_HEADERS[4])
-
-    def select_timespan(self, option: str) -> None:
-        element = self.wait_to_be_visible(self.Locator.TIMESPAN_DROPDOWN, timeout=5)
-        Select(element).select_by_visible_text(option)
-
-    def get_graph_for_selected_timespan_option(self, option: str) -> WebElement | bool:
-        self.select_timespan(option)
-        graph_locator = self.Locator.memory_usage_graph(SI.TIMESPAN_OPTIONS.get(option))
-        try:
-            return self.wait_to_be_visible(graph_locator, timeout=5)
-        except TimeoutException:
-            return False
 
     def click_on_thread_dump_page_link(self) -> Self:
         self.click_on(self.Locator.THREAD_DUMP_PAGE_LINK)
