+++ tests/manage_jenkins/data.py
@@ -1,3 +1,7 @@
+class ManageJenkinsData:
+    TIMESPAN_OPTIONS = {"Short": "sec10&width", "Medium": "min&width", "Long": "hour&width"}
+
+
 class SystemInformationData:
     TABS_BAR_HEADERS = [
         "System Properties",
@@ -7,4 +11,3 @@ class SystemInformationData:
         "Thread Dumps"
     ]
     SHOW_SINGLE_VALUE_BUTTON_TEXT = "Hidden value, click to show this value"
-    TIMESPAN_OPTIONS = {"Short": "sec10&width", "Medium": "min&width", "Long": "hour&width"}
