+++ tests/manage_jenkins/status_information/test_load_statistics.py
@@ -1,3 +1,19 @@
+import allure
+from tests.manage_jenkins.data import ManageJenkinsData as MJ
+
+
 def test_open_load_statistics_page(manage_jenkins_page):
     load_stat_page = manage_jenkins_page.go_to_load_statistics_page()
     assert load_stat_page.get_title() == "Jenkins Load Statistics [Jenkins]", "Load statistics page title is incorrect"
+
+
+@allure.epic("Manage Jenkins")
+@allure.story("Load Statistics")
+@allure.title("Display graph according to the selected Timespan")
+@allure.description("Verify that the load statistics graph is displayed according to the selected timespan.")
+@allure.testcase("TC_10.022.02")
+@allure.link("https://github.com/RedRoverSchool/JenkinsQA_Python_2025_spring/issues/722", name="Github issue")
+def test_load_statistics_page_select_timespan(load_statistics_page):
+    for option in MJ.TIMESPAN_OPTIONS:
+        assert load_statistics_page.get_graph_for_selected_timespan_option(option), \
+            f"Graph for timespan option '{option}' is not displayed, or displayed incorrectly"
